fix(test): Reset database before setup(grant)
