+++ src/test/kotlin/me/loghub/api/controller/series/SeriesReviewControllerTest.kt
@@ -4,51 +4,48 @@ import me.loghub.api.dto.auth.token.TokenDTO
 import me.loghub.api.dto.series.review.PostSeriesReviewDTO
 import me.loghub.api.entity.user.User
 import me.loghub.api.service.test.TestGrantService
+import me.loghub.api.util.resetDatabase
 import org.junit.jupiter.api.*
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.web.client.TestRestTemplate
-import org.springframework.core.io.ClassPathResource
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
 import org.springframework.http.RequestEntity
 import org.springframework.http.ResponseEntity
 import org.springframework.jdbc.core.JdbcTemplate
-import org.springframework.jdbc.datasource.init.DatabasePopulatorUtils
-import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator
 import org.springframework.test.context.ActiveProfiles
 import kotlin.test.assertEquals
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @TestClassOrder(ClassOrderer.OrderAnnotation::class)
 @ActiveProfiles("test")
 class SeriesReviewControllerTest(
     @Autowired private val rest: TestRestTemplate,
     @Autowired private val jdbcTemplate: JdbcTemplate,
 ) {
-    companion object {
-        lateinit var member1: User
-        lateinit var member1Token: TokenDTO
-        lateinit var member2: User
-        lateinit var member2Token: TokenDTO
+    lateinit var member1: User
+    lateinit var member1Token: TokenDTO
+    lateinit var member2: User
+    lateinit var member2Token: TokenDTO
 
-        object SeriesReview {
-            object Id {
-                const val BY_MEMBER1 = 1L
-                const val INVALID = 999L
-            }
+    object SeriesReview {
+        object Id {
+            const val BY_MEMBER1 = 1L
+            const val INVALID = 999L
         }
+    }
 
-        @JvmStatic
-        @BeforeAll
-        fun setup(@Autowired grantService: TestGrantService) {
-            val (member1, member1Token) = grantService.grant("member1")
-            this.member1 = member1
-            this.member1Token = member1Token
-            val (member2, member2Token) = grantService.grant("member2")
-            this.member2 = member2
-            this.member2Token = member2Token
-        }
+    @BeforeAll
+    fun setup(@Autowired grantService: TestGrantService) {
+        resetDatabase(jdbcTemplate)
+        val (member1, member1Token) = grantService.grant("member1")
+        this.member1 = member1
+        this.member1Token = member1Token
+        val (member2, member2Token) = grantService.grant("member2")
+        this.member2 = member2
+        this.member2Token = member2Token
     }
 
     private inline fun <reified T> getReviews(seriesId: Long) =
@@ -85,23 +82,12 @@ class SeriesReviewControllerTest(
         return rest.exchange(request.build(), T::class.java)
     }
 
-    private fun resetDatabase() {
-        val dataSource = jdbcTemplate.dataSource
-            ?: error("DataSource is required for resetting database")
-        val populator = ResourceDatabasePopulator().apply {
-            addScript(ClassPathResource("/database/data/truncate.sql"))
-            addScript(ClassPathResource("/database/data/test.sql"))
-        }
-
-        DatabasePopulatorUtils.execute(populator, dataSource)
-    }
-
     @Nested
     @Order(1)
     @TestInstance(TestInstance.Lifecycle.PER_CLASS)
     inner class GetReviews {
         @BeforeAll
-        fun setupDatabase() = resetDatabase()
+        fun setupDatabase() = resetDatabase(jdbcTemplate)
 
         @Test
         fun `getReviews - ok`() {
@@ -121,7 +107,7 @@ class SeriesReviewControllerTest(
         )
 
         @BeforeAll
-        fun setupDatabase() = resetDatabase()
+        fun setupDatabase() = resetDatabase(jdbcTemplate)
 
         @Test
         fun `postReview - unauthenticated`() {
@@ -146,7 +132,7 @@ class SeriesReviewControllerTest(
         )
 
         @BeforeAll
-        fun setupDatabase() = resetDatabase()
+        fun setupDatabase() = resetDatabase(jdbcTemplate)
 
         @Test
         fun `editReview - unauthenticated`() {
@@ -179,7 +165,7 @@ class SeriesReviewControllerTest(
     @TestMethodOrder(MethodOrderer.OrderAnnotation::class)
     inner class DeleteReview {
         @BeforeAll
-        fun setupDatabase() = resetDatabase()
+        fun setupDatabase() = resetDatabase(jdbcTemplate)
 
         @Test
         fun `deleteReview - unauthenticated`() {
